FEM: unit test, move rename fem common test module
