+++ src/Mod/Fem/CMakeLists.txt
@@ -111,7 +111,7 @@ SET(FemSolverZ88_SRCS
 SET(FemTests_SRCS
     femtest/__init__.py
     femtest/testccxtools.py
-    femtest/testfemcommon.py
+    femtest/testcommon.py
     femtest/testsolverframework.py
     femtest/testtools.py
 )
